# rawTileData cleanup