--- a/bench/benchmarks/buffer.js
+++ b/bench/benchmarks/buffer.js
@@ -158,7 +158,7 @@ function runSample(stylesheet, getGlyphs, getIcons, getTile, callback) {
         getTile(url, function(err, response) {
             if (err) throw err;
             var data = new VT.VectorTile(new Protobuf(new Uint8Array(response)));
-            workerTile.parse(data, layerFamilies, actor, null, function(err) {
+            workerTile.parse(data, layerFamilies, actor, function(err) {
                 if (err) return callback(err);
                 eachCallback();
             });