--- a/bench/benchmarks/buffer.js
+++ b/bench/benchmarks/buffer.js
@@ -157,7 +157,7 @@ function runSample(stylesheet, getGlyphs, getIcons, getTile, callback) {
 
         getTile(url, function(err, response) {
             if (err) throw err;
-            var data = new VT.VectorTile(new Protobuf(new Uint8Array(response)));
+            var data = new VT.VectorTile(new Protobuf(response));
             workerTile.parse(data, layerFamilies, actor, function(err) {
                 if (err) return callback(err);
                 eachCallback();