--- a/bin/point-in-polygon.js
+++ b/bin/point-in-polygon.js
@@ -22,7 +22,7 @@ var data = fs.readFileSync(process.argv[2]);
 zlib.inflate(data, function(err, data) {
 if (err) throw err;
 
-var tile = new VectorTile(new Protobuf(new Uint8Array(data)));
+var tile = new VectorTile(new Protobuf(data));
 
 
 var tree = rbush(9, ['.x1', '.y1', '.x2', '.y2']);