--- a/bin/show.js
+++ b/bin/show.js
@@ -29,7 +29,7 @@ var data = fs.readFileSync(process.argv[2]);
 zlib.inflate(data, function(err, data) {
     if (err) throw err;
 
-    var tile = new VectorTile(new Protobuf(new Uint8Array(data)));
+    var tile = new VectorTile(new Protobuf(data));
 
     for (var layer_name in tile.layers) {
         var layer = tile.layers[layer_name];