--- a/js/symbol/glyph_source.js
+++ b/js/symbol/glyph_source.js
@@ -106,7 +106,7 @@ GlyphSource.prototype.loadRange = function(fontstack, range, callback) {
         var url = glyphUrl(fontstack, rangeName, this.url);
 
         getArrayBuffer(url, function(err, data) {
-            var glyphs = !err && new Glyphs(new Protobuf(new Uint8Array(data)));
+            var glyphs = !err && new Glyphs(new Protobuf(data));
             for (var i = 0; i < loading[range].length; i++) {
                 loading[range][i](err, range, glyphs);
             }