--- a/test/js/data/load_geometry.test.js
+++ b/test/js/data/load_geometry.test.js
@@ -8,7 +8,7 @@ var VectorTile = require('vector-tile').VectorTile;
 var loadGeometry = require('../../../js/data/load_geometry.js');
 
 // Load a line feature from fixture tile.
-var vt = new VectorTile(new Protobuf(new Uint8Array(fs.readFileSync(path.join(__dirname, '/../../fixtures/mbsv5-6-18-23.vector.pbf')))));
+var vt = new VectorTile(new Protobuf(fs.readFileSync(path.join(__dirname, '/../../fixtures/mbsv5-6-18-23.vector.pbf'))));
 
 test('loadGeometry', function(t) {
     var feature = vt.layers.road.feature(0);