--- a/test/js/util/classify_rings.test.js
+++ b/test/js/util/classify_rings.test.js
@@ -8,7 +8,7 @@ var VectorTile = require('vector-tile').VectorTile;
 var classifyRings = require('../../../js/util/classify_rings');
 
 // Load a fill feature from fixture tile.
-var vt = new VectorTile(new Protobuf(new Uint8Array(fs.readFileSync(path.join(__dirname, '/../../fixtures/mbsv5-6-18-23.vector.pbf')))));
+var vt = new VectorTile(new Protobuf(fs.readFileSync(path.join(__dirname, '/../../fixtures/mbsv5-6-18-23.vector.pbf'))));
 var feature = vt.layers.water.feature(0);
 
 test('classifyRings', function(assert) {