<!--
Hello! Thanks for contributing. To help your PR be most easily reviewed, please complete
the following checklist:
-->

In preparation for https://github.com/mapbox/mapbox-gl-native/issues/5701, I reviewed the GL JS implementation of tile reloading and noticed a few issues.

* Whenever a `VectorTile` is created, we were accidentally making an extra copy of the data.
* This was hiding a latent bug, where the worker thread was giving up ownership of raw data that it needed to retain.
* We were copying raw data _back_ to the worker on tile reload requests for no reason.
* The naming of tile and data related variables in `VectorTileWorkerSource` was vague and confusing.
* The contract of the semi-public `loadVectorData` extension point was underspecified.

This PR fixes those issues.

## Benchmarks

### `master`

**load-multiple-maps:** 150 ms, loaded 6 maps.
**buffer:** 854 ms
**fps:** 60 fps
**frame-duration:** 7.6 ms, 0% > 16ms
**query-point:** 0.63 ms
**query-box:** 51.76 ms
**geojson-setdata-small:** 11 ms
**geojson-setdata-large:** 97 ms

### Branch

**load-multiple-maps:** 151 ms, loaded 6 maps.
**buffer:** 863 ms
**fps:** 60 fps
**frame-duration:** 7.6 ms, 0% > 16ms
**query-point:** 0.65 ms
**query-box:** 49.49 ms
**geojson-setdata-small:** 10 ms
**geojson-setdata-large:** 104 ms

## Checklist

 - [x] briefly describe the changes in this PR
 - [x] [post scores for all benchmarks on your branch and `master`](https://github.com/mapbox/mapbox-gl-js/blob/master/bench/README.md#running-benchmarks)
 - [x] [do a quick sanity check on the debug page](https://github.com/mapbox/mapbox-gl-js/blob/master/CONTRIBUTING.md#serving-the-debug-page)
 - [ ] get a PR review :+1: / :-1: