@mourner 2016-08-26
Looks great to me overall! Are we sure the main thread isn't blocked when `rawTileData` is transferred with a copy? Or does this only happen in the other direction, when the main thread sends to the worker?